+++ NEWS.md
@@ -1,4 +1,10 @@
 
+# 1.0.0.9000
+
+* Add empty directories to the archive
+
+* Include directories in the file list
+
 # 1.0.0
 
 First public release.